show email form on download?start=true
