+++ src/containers/Downloads/index.js
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from 'react';
 import Helmet from 'react-helmet';
 import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
+import { Link, Redirect } from 'react-router-dom';
 import BackToTop from '../../components/BackToTop';
 import DownloadBar from './DownloadBar';
 import DownloadDetails from './DownloadDetails';
@@ -9,10 +9,24 @@ import './Downloads.scss';
 import NoDatasetsImage from './../../common/images/no-datasets.svg';
 import Loader from '../../components/Loader';
 import { fetchDataSetDetails } from '../../state/download/actions';
+import { getQueryParamObject } from '../../common/helpers';
+import DownloadStart from './DownloadStart/DownloadStart';
 
 class Download extends Component {
   render() {
-    const { dataSetId, dataSet, aggregate_by, scale_by } = this.props;
+    const { dataSetId, dataSet, aggregate_by, scale_by, location } = this.props;
+    const dataSetCanBeDownloaded = dataSet && Object.keys(dataSet).length > 0;
+    const params = getQueryParamObject(location.search);
+
+    // show form to get information and start the download
+    if (params.start === 'true') {
+      if (dataSetCanBeDownloaded) {
+        return <DownloadStart dataSetId={dataSetId} />;
+      } else {
+        // if the dataset can't be downloaded, go back to the downloads page.
+        return <Redirect to="/download" />;
+      }
+    }
 
     return (
       <div className="downloads">
@@ -25,7 +39,7 @@ class Download extends Component {
           {({ isLoading }) =>
             isLoading ? (
               <div className="loader" />
-            ) : !dataSet || !Object.keys(dataSet).length ? (
+            ) : !dataSetCanBeDownloaded ? (
               <div className="downloads__empty">
                 <h3 className="downloads__empty-heading">
                   Your dataset is empty.
